requirements.txt
annotated-types==0.5.0
anyio==4.0.0
ariadne==0.20.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
Flask==2.3.3
graphql-core==3.2.3
graphqlclient==0.2.4
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
prisma==0.10.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.31.1
tomlkit==0.12.1
typing_extensions==4.8.0
uvicorn==0.23.2
Werkzeug==2.3.7
